Rename OCR test_utils to clarify content
