+++ ocr/ocr__accuracy_test_utils.py
